pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scalable-moe-lora"
version = "0.1.0"
description = "Scalable MoE-LoRA: a controlled factorial study of MoE adapter design axes on a frozen LLaMA 3.2 1B."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
authors = [
    { name = "Anh Trung Nguyen" },
    { name = "Jiaqi Lu" },
    { name = "Andrew Huang" },
]
keywords = ["lora", "moe", "mixture-of-experts", "llama", "fine-tuning", "peft"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.1",
    "transformers>=4.40",
    "datasets>=2.14",
    "accelerate>=0.25",
    "evaluate>=0.4",
    "rouge-score>=0.1",
    "nltk>=3.8",
    "PyYAML>=6.0",
    "tqdm>=4.65",
    "numpy>=1.24",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1",
]

[project.scripts]
moe-lora-train = "scalable_moe_lora.train_reasoning:main"
moe-lora-eval = "scalable_moe_lora.evaluate_reasoning:main"
moe-lora-routing = "scalable_moe_lora.analysis.per_layer_routing:main"
moe-lora-gates = "scalable_moe_lora.analysis.gate_magnitudes:main"
moe-lora-correctness = "scalable_moe_lora.analysis.correctness:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
scalable_moe_lora = ["py.typed"]
